+++ changelog.md
@@ -1,3 +1,11 @@
+## v1.0.2
+
+- Added **Discord Button** on the main menu
+- Added **Badge Descriptions** when clicking on the badges on the profile page or comment cells
+- Added **Promote User** for Gauntlet Managers
+- Improved the **Gauntlet Info Popup** with more information about the mod and its features
+- Fixed the **Gauntlet Role Badge** scaling on comment cells
+
 ## v1.0.1
 
 - Massive blunder. Fixed the crash when clicking the sync on mobile.